% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal.R
\docType{data}
\name{legal}
\alias{legal}
\title{legal}
\format{
dataframe com uma coluna:
\describe{
\item{\code{word}}{termos comuns}
}
}
\source{
\url{https://tjsp.jus.br}
}
\usage{
legal
}
\description{
Palavras comuns do vocabulário legal
}
\keyword{datasets}
