% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logradouros.R
\docType{data}
\name{logradouros}
\alias{logradouros}
\title{logradouros}
\format{
dataframe com uma coluna:
\describe{
\item{\code{word}}{tipo de logradouro}
}
}
\source{
\url{https://gist.github.com/jjesusfilho/10524df296794b2f4673b42b4f8c8c3a}
}
\usage{
logradouros
}
\description{
Tipos de logradouros
}
\keyword{datasets}
