% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magistrados.R
\docType{data}
\name{magistrados}
\alias{magistrados}
\title{magistrados}
\format{
dataframe com duas colunas:
\describe{
\item{\code{tribunal}}{sigla do tribunal}
\item{\code{word}}{nome ou sobrenome}
}
}
\source{
\url{https://www.cnj.jus.br/corregedoria/justica_aberta/?}
}
\usage{
magistrados
}
\description{
nomes e sobrenomes dos magistrados de
todos os tribunais estaduais e federais
}
\keyword{datasets}
