% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomes.R
\docType{data}
\name{nomes}
\alias{nomes}
\title{nomes}
\format{
dataframe com uma coluna:
\describe{
\item{\code{word}}{primeiro nome de pessoas}
}
}
\source{
\url{https://censo2010.ibge.gov.br/nomes/}
}
\usage{
nomes
}
\description{
Nomes obtidos da base do IBGE
}
\keyword{datasets}
