% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pessoas.R
\docType{data}
\name{pessoas}
\alias{pessoas}
\title{pessoas}
\format{
dataframe com uma coluna:
\describe{
\item{\code{word}}{pessoa do processo}
}
}
\source{
\url{https://tjsp.jus.br}
}
\usage{
pessoas
}
\description{
Pessoas do processo
}
\keyword{datasets}
