% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portugues.R
\docType{data}
\name{portugues}
\alias{portugues}
\title{portugues}
\format{
dataframe com uma coluna:
\describe{
\item{\code{word}}{palavra}
}
}
\source{
\url{https://github.com/andresionek91/portuguese-stop-words}
}
\usage{
portugues
}
\description{
Stopwords mais comuns do Português
}
\keyword{datasets}
