% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufs.R
\docType{data}
\name{ufs}
\alias{ufs}
\title{ufs}
\format{
dataframe com duas colunas:
\describe{
\item{\code{uf}}{unidade federativa}
\item{\code{sigla}}{sigla da unidade federativa}
}
}
\source{
\url{https://ibge.gov.br}
}
\usage{
ufs
}
\description{
Unidades federativas
}
\keyword{datasets}
